+++ playground/src/app/page.tsx
@@ -1,107 +1,57 @@
-// import {
-// basehub,
-// RichTextJson,
-// TweetComponentGenqlSelection,
-// FieldsSelection,
-// } from "basehub";
-// import { RichText } from "basehub/react";
-// import { Tweet } from "react-tweet";
-import { basehub } from "@basehub";
+import { Pump } from "../../.basehub/next-pump";
 
 export default async function HomePage() {
-  // const data = await basehub({ draft: true }).raw({
-  //   query: `{ __typename, testBreak {
-  //     authors {
-  //       items {
-  //         _slug
-  //         _title
-  //       }
-  //     }
-  //   } }`,
-  // });
-
-  const otherData = await basehub({
-    cache: "no-store",
-    // draft: true,
-  }).query({
-    // __name: true,
-    // _sys: {
-    //   slug: true,
-    // },
-    homepage: {
-      _id: true,
-    },
-    // docsud: {
-    //   _id: true,
-    // },
-    // testBreak: {
-    //   authors: {
-    //     items: {
-    //       _slug: true,
-    //       _title: true,
-    //     },
-    //   },
-    // },
-    // blogIndex: {
-    //   blogIndex: true,
-    //   featuredPost: {
-    //     content: {
-    //       json: {
-    //         toc: true,
-    //         content: true,
-    //         blocks: {
-    //           __typename: true,
-    //           on_BoldTextComponent: {
-    //             _title: true,
-    //             _sys: {
-    //               id: true,
-    //             },
-    //           },
-    //           on_TweetComponent: {
-    //             _sys: {
-    //               id: true,
-    //             },
-    //             tweetId: true,
-    //           },
-    //         },
-    //       },
-    //     },
-    //   },
-    // },
-  });
-
   return (
     <main>
-      <pre>{JSON.stringify({ otherData }, null, 2)}</pre>
-      <br />
-      {/* <RichText
-        blocks={blogIndex.featuredPost.content?.json.blocks}
-        components={{
-          p: (props) => <p {...props} />,
-          code: ({ children, isInline }) => (
-            <code style={{ color: isInline ? "red" : "green" }}>
-              {children}
-            </code>
-          ),
-          TweetComponent: ({ tweetId }) => {
-            return <Tweet id={tweetId} />;
+      <h1>Does it work?</h1>
+      <Pump
+        draft
+        token={process.env.BASEHUB_TOKEN!}
+        queries={[
+          {
+            doc: {
+              heroTitle: true,
+              _title: true,
+            },
           },
-          video: ({ src }) => {
-            return <video src={src} autoPlay loop playsInline muted />;
-          },
-        }}
+        ]}
       >
-        {blogIndex.featuredPost.content?.json.content}
-      </RichText>
+        {async ([data]) => {
+          "use server";
 
-      <RichText
-        components={{
-          ol: (props) => <ol {...props} style={{ color: "red !important" }} />,
-          a: (props) => <a {...props} style={{ color: "green !important" }} />,
+          if (!data) return null;
+          return <h1>{JSON.stringify(data, null, 2)}</h1>;
         }}
+      </Pump>
+      <Pump
+        draft
+        token={process.env.BASEHUB_TOKEN!}
+        queries={[
+          {
+            doc: {
+              heroTitle: true,
+              _title: true,
+            },
+          },
+          {
+            _sys: {
+              hash: true,
+            },
+          },
+        ]}
       >
-        {blogIndex.featuredPost.content?.json.toc}
-      </RichText> */}
+        {async ([data, data2]) => {
+          "use server";
+
+          if (!data) return null;
+          return (
+            <h1>
+              different return data even...{data.doc.heroTitle}{" "}
+              {data2._sys.hash}
+            </h1>
+          );
+        }}
+      </Pump>
     </main>
   );
 }